Split multi and single bodies into their own comps.
