+++ src/applications/check-in/travel-claim/pages/travel-mileage/index.jsx
@@ -4,8 +4,8 @@ import { useTranslation, Trans } from 'react-i18next';
 import PropTypes from 'prop-types';
 // eslint-disable-next-line import/no-unresolved
 import { recordEvent } from '@department-of-veterans-affairs/platform-monitoring/exports';
-import { VaCheckbox } from '@department-of-veterans-affairs/component-library/dist/react-bindings';
 
+import { useFormRouting } from '../../../hooks/useFormRouting';
 import { createAnalyticsSlug } from '../../../utils/analytics';
 import { setFacilityToFile } from '../../../actions/travel-claim';
 import { hasMultipleFacilities } from '../../../utils/appointment';
@@ -13,8 +13,8 @@ import { makeSelectCurrentContext, makeSelectForm } from '../../../selectors';
 import { APP_NAMES } from '../../../utils/appConstants';
 import Wrapper from '../../../components/layout/Wrapper';
 import BackButton from '../../../components/BackButton';
-
-import { useFormRouting } from '../../../hooks/useFormRouting';
+import SingleFacilityBody from './SingleFacilityBody';
+import MultipleFacilityBody from './MultiFacilityBody';
 
 const TravelMileage = props => {
   const { router } = props;
@@ -78,29 +78,6 @@ const TravelMileage = props => {
     },
     [dispatch, goToNextPage, selectedFacilities, setError],
   );
-  const onCheck = useCallback(
-    e => {
-      const stationNo = e.target.value;
-      const { checked } = e.detail;
-      const appointments = appointmentsByFacility[stationNo];
-      const firstAppointment = appointments[0];
-      const value = {
-        stationNo: firstAppointment.stationNo,
-        startTime: firstAppointment.startTime,
-        multipleAppointments: appointments.length > 1,
-      };
-      let newFacilities;
-      if (checked) {
-        newFacilities = [...selectedFacilities, value];
-      } else {
-        newFacilities = selectedFacilities.filter(
-          fac => fac.stationNo !== value.stationNo,
-        );
-      }
-      setSelectedFacilities(newFacilities);
-    },
-    [appointmentsByFacility, selectedFacilities],
-  );
   const formatAppointment = appointment => {
     const appointmentLabel = appointment.clinicStopCodeName
       ? `${appointment.clinicStopCodeName} ${t('appointment')}`
@@ -112,27 +89,6 @@ const TravelMileage = props => {
   let header = t('file-mileage-only-claim-todays-appointment', {
     count: eligibleToFile.length,
   });
-  let body = (
-    <div
-      className="vads-u-border-top--1px vads-u-border-bottom--1px vads-u-padding-y--1 vads-u-margin-y--2 vads-u-border-color--gray-light"
-      data-testid="single-fac-context"
-    >
-      <span className="vads-u-font-weight--bold">
-        {eligibleToFile[0].facility}
-      </span>
-      <div className="vads-u-margin-y--1">
-        {eligibleToFile.map(appointment => (
-          <p
-            className="vads-u-margin--0"
-            key={appointment.appointmentIen}
-            data-testid={`appointment-list-item-${appointment.appointmentIen}`}
-          >
-            {formatAppointment(appointment)}
-          </p>
-        ))}
-      </div>
-    </div>
-  );
   if (multipleFacilities) {
     appointmentsByFacility = {};
     eligibleToFile.forEach(appointment => {
@@ -143,35 +99,6 @@ const TravelMileage = props => {
       }
     });
     header = t('select-appointments-to-file-today');
-    body = (
-      <div data-testid="multi-fac-context">
-        <p>{t('if-youre-filing-only-mileage-no-other-file-all-claims-now')}</p>
-        <va-checkbox-group
-          data-testid="checkbox-group"
-          error={error ? t('select-one-or-more-appointments') : ''}
-          uswds
-          class="vads-u-margin-top--0 vads-u-margin-bottom--4"
-        >
-          {Object.keys(appointmentsByFacility).map(facility => (
-            <VaCheckbox
-              data-testid={`checkbox-${facility}`}
-              key={facility}
-              uswds
-              tile
-              value={facility}
-              label={appointmentsByFacility[facility][0].facility}
-              checkbox-description={appointmentsByFacility[facility].map(
-                appointment => ` ${formatAppointment(appointment)}`,
-              )}
-              onVaChange={onCheck}
-              checked={selectedFacilities.some(
-                fac => fac.stationNo === facility,
-              )}
-            />
-          ))}
-        </va-checkbox-group>
-      </div>
-    );
   }
   return (
     <>
@@ -183,7 +110,21 @@ const TravelMileage = props => {
       <Wrapper pageTitle={header} classNames="travel-page" withBackButton>
         {/* Setting state value here for testing purposes. Could not mock hook with our test setup. */}
         <div data-testid={JSON.stringify(selectedFacilities)}>
-          {body}
+          {multipleFacilities ? (
+            <MultipleFacilityBody
+              error={error}
+              formatAppointment={formatAppointment}
+              appointmentsByFacility={appointmentsByFacility}
+              selectedFacilities={selectedFacilities}
+              setSelectedFacilities={setSelectedFacilities}
+            />
+          ) : (
+            <SingleFacilityBody
+              facility={eligibleToFile[0].facility}
+              appointments={eligibleToFile}
+              formatAppointment={formatAppointment}
+            />
+          )}
           <va-additional-info
             trigger={t('if-you-have-other-expenses-to-claim')}
             uswds
